server/requirements.txt
flask
firebase-admin 
supabase
python-dotenv